feat: text component caption variant
