+++ src/app/shared/components/template/components/text/text.component.html
@@ -1,0 +1,9 @@
+<div
+  *ngIf="_row && _row.value"
+  class="large standard normal margin-t-large"
+  [class]="params.style"
+  [ngStyle]="isFalsy ? { display: 'none' } : { display: 'block' }"
+  [innerHTML]="type === 'numbered' ? (_row.value | number) : (_row.value?.toString() | markdown)"
+  [style.textAlign]="params.textAlign"
+  [attr.data-variant]="params.variant"
+></div>
